+++ sandbox/src/test/kotlin/com/bumble/appyx/sandbox/client/mvicoreexample/leaf/MviCoreLeafNodeTest.kt
@@ -4,6 +4,7 @@ import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Lifecycle
 import com.bumble.appyx.core.modality.BuildContext
 import com.bumble.appyx.core.node.Node
+import com.bumble.appyx.interop.rx2.plugin.disposeOnDestroyPlugin
 import com.bumble.appyx.sandbox.client.mvicoreexample.MviCoreExampleNode.NavTarget
 import com.bumble.appyx.sandbox.client.mvicoreexample.MviCoreExampleViewImpl.Event
 import com.bumble.appyx.sandbox.client.mvicoreexample.feature.MviCoreExampleFeature.News
@@ -15,6 +16,7 @@ import com.bumble.appyx.sandbox.stub.NodeViewStub
 import com.bumble.appyx.testing.junit4.util.MainDispatcherRule
 import com.bumble.appyx.testing.unit.common.helper.NodeTestHelper
 import com.bumble.appyx.testing.unit.common.helper.nodeTestHelper
+import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
@@ -47,7 +49,7 @@ class MviCoreLeafNodeTest {
         node = MviCoreLeafNode(
             buildContext = BuildContext.root(savedStateMap = null),
             view = view,
-            interactor = interactor,
+            plugins = listOf(interactor, disposeOnDestroyPlugin(feature)),
         )
 
         testHelper = node.nodeTestHelper()
@@ -71,5 +73,11 @@ class MviCoreLeafNodeTest {
 
             testObserver.assertValues(ViewModel.InitialState(stateName), ViewModel.Loading)
         }
+    }
+
+    @Test
+    fun `when state is destroyed then feature is disposed`() {
+        testHelper.moveTo(Lifecycle.State.DESTROYED)
+        assertTrue(feature.isDisposed)
     }
 }